Use multiple C* workers to restore() with unlimited queue
This solution uses unlimited queue for output (read from C*) blocks
to not block ouput file writer if blocks are not read in expected order.

It has a disadvantage of potentially growing queue of read blocks.

Experiments show that with 480M file and 4 readers the queue max is 50 blocks
with limited block size to 64K it makes 3M of memory.